# nodejs-server-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is Swagger UI environment generated for the TMF Service Catalog\
    \ Management specification"
  version: "2.0.0"
  title: "Service Catalog Management"
host: "localhost:8080"
basePath: "/tmf-api/serviceCatalogManagement/v2/"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /serviceCatalog:
    get:
      tags:
      - "serviceCatalog"
      summary: "List or find 'ServiceCatalog' objects"
      operationId: "listServiceCatalog"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma separated properties to display in response"
        required: false
        type: "string"
      - name: "name"
        in: "query"
        description: "For filtering: Name of the catalog"
        required: false
        type: "string"
      - name: "type"
        in: "query"
        description: "For filtering: Indicates the (class) type of catalog. For service\
          \ catalogs, this will be 'ServiceCatalog'."
        required: false
        type: "string"
      - name: "@schemaLocation"
        in: "query"
        description: "For filtering: This field provides a link to the schema describing\
          \ this REST resource"
        required: false
        type: "string"
      - name: "@baseType"
        in: "query"
        description: "For filtering: Indicates<b> </b>the base (class) type of this\
          \ REST resource"
        required: false
        type: "string"
      - name: "version"
        in: "query"
        description: "For filtering: Catalog version"
        required: false
        type: "string"
      - name: "validForStartDateTime"
        in: "query"
        description: "For filtering: An instant of time, starting at the TimePeriod"
        required: false
        type: "string"
        format: "date-time"
      - name: "validForEndDateTime"
        in: "query"
        description: "For filtering: An instant of time, ending at the TimePeriod."
        required: false
        type: "string"
        format: "date-time"
      - name: "lastUpdate"
        in: "query"
        description: "For filtering: Date and time of the last update"
        required: false
        type: "string"
        format: "date-time"
      - name: "lifecycleStatus"
        in: "query"
        description: "For filtering: Used to indicate the current lifecycle status"
        required: false
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceCatalog"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCatalog"
    post:
      tags:
      - "serviceCatalog"
      summary: "Creates a 'ServiceCatalog'"
      operationId: "createServiceCatalog"
      parameters:
      - in: "body"
        name: "serviceCatalog"
        description: "The Service Catalog to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceCatalog_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceCatalog"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCatalog"
  /serviceCatalog/{id}:
    get:
      tags:
      - "serviceCatalog"
      summary: "Retrieves a 'ServiceCatalog' by Id"
      operationId: "retrieveServiceCatalog"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Catalog"
        required: true
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceCatalog"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCatalog"
    delete:
      tags:
      - "serviceCatalog"
      summary: "Deletes a 'ServiceCatalog' by Id"
      operationId: "deleteServiceCatalog"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Catalog"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCatalog"
    patch:
      tags:
      - "serviceCatalog"
      summary: "Updates partially a 'ServiceCatalog' by Id"
      operationId: "patchServiceCatalog"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Catalog"
        required: true
        type: "string"
      - in: "body"
        name: "serviceCatalog"
        description: "The Service Catalog to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceCatalog_Update"
      responses:
        201:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceCatalog"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCatalog"
  /serviceCategory:
    get:
      tags:
      - "serviceCategory"
      summary: "List or find 'ServiceCategory' objects"
      operationId: "listServiceCategory"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma separated properties to display in response"
        required: false
        type: "string"
      - name: "name"
        in: "query"
        description: "For filtering: Name of the category"
        required: false
        type: "string"
      - name: "type"
        in: "query"
        description: "For filtering: The (class) type of this category"
        required: false
        type: "string"
      - name: "@schemalLocation"
        in: "query"
        description: "For filtering: This field provides a link to the schema describing\
          \ this REST resource"
        required: false
        type: "string"
      - name: "@baseType"
        in: "query"
        description: "For filtering: Immediate base class type of this category"
        required: false
        type: "string"
      - name: "version"
        in: "query"
        description: "For filtering: Category version"
        required: false
        type: "string"
      - name: "validForStartDateTime"
        in: "query"
        description: "For filtering: An instant of time, starting at the TimePeriod"
        required: false
        type: "string"
        format: "date-time"
      - name: "validForEndDateTime"
        in: "query"
        description: "For filtering: An instant of time, ending at the TimePeriod."
        required: false
        type: "string"
        format: "date-time"
      - name: "lifecycleStatus"
        in: "query"
        description: "For filtering: Used to indicate the current lifecycle status"
        required: false
        type: "string"
      - name: "lastUpdate"
        in: "query"
        description: "For filtering: Date and time of the last update"
        required: false
        type: "string"
        format: "date-time"
      - name: "parentId"
        in: "query"
        description: "For filtering: Unique identifier of the parent category"
        required: false
        type: "string"
      - name: "isRoot"
        in: "query"
        description: "For filtering: If true, this Boolean indicates that the category\
          \ is a root of categories"
        required: false
        type: "boolean"
      - name: "relatedPartyRole"
        in: "query"
        description: "For filtering: Role of the related party."
        required: false
        type: "string"
      - name: "relatedPartyName"
        in: "query"
        description: "For filtering: Name of the related party"
        required: false
        type: "string"
      - name: "serviceCandidateVersion"
        in: "query"
        description: "For filtering: ServiceCandidate version"
        required: false
        type: "string"
      - name: "serviceCandidateName"
        in: "query"
        description: "For filtering: Name given to the ServiceCandidate"
        required: false
        type: "string"
      - name: "serviceCandidateType"
        in: "query"
        description: "For filtering: The (class) type of the ServiceCandidate"
        required: false
        type: "string"
      - name: "categoryVersion"
        in: "query"
        description: "For filtering: Category version"
        required: false
        type: "string"
      - name: "categoryName"
        in: "query"
        description: "For filtering: Name of the category"
        required: false
        type: "string"
      - name: "categoryType"
        in: "query"
        description: "For filtering: the class type of the category"
        required: false
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceCategory"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCategory"
    post:
      tags:
      - "serviceCategory"
      summary: "Creates a 'ServiceCategory'"
      operationId: "createServiceCategory"
      parameters:
      - in: "body"
        name: "serviceCategory"
        description: "The Service Category to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceCategory_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceCategory"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCategory"
  /serviceCategory/{id}:
    get:
      tags:
      - "serviceCategory"
      summary: "Retrieves a 'ServiceCategory' by Id"
      operationId: "retrieveServiceCategory"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Category"
        required: true
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceCategory"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCategory"
    delete:
      tags:
      - "serviceCategory"
      summary: "Deletes a 'ServiceCategory' by Id"
      operationId: "deleteServiceCategory"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Category"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCategory"
    patch:
      tags:
      - "serviceCategory"
      summary: "Updates partially a 'ServiceCategory' by Id"
      operationId: "patchServiceCategory"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Category"
        required: true
        type: "string"
      - in: "body"
        name: "serviceCategory"
        description: "The Service Category to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceCategory_Update"
      responses:
        201:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceCategory"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCategory"
...
definitions:
  ServiceCatalog:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the Catalog"
      href:
        type: "string"
        description: "Unique reference of the catalog"
      name:
        type: "string"
        description: "Name of the catalog"
      description:
        type: "string"
        description: "Description of this catalog"
      '@type':
        type: "string"
        description: "Indicates the (class) type of catalog. For service catalogs,\
          \ this will be 'ServiceCatalog'."
      '@schemaLocation':
        type: "string"
        description: "This field provides a link to the schema describing this REST\
          \ resource"
      '@baseType':
        type: "string"
        description: "Indicates<b> </b>the base (class) type of this REST resource"
      version:
        type: "string"
        description: "Catalog version"
      validFor:
        description: "The period for which the catalog is valid"
        $ref: "#/definitions/TimePeriod"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status"
    description: "The root entity for service catalog management.\nA service catalog\
      \ is a group of service specifications made available through service candidates\
      \ that an organization provides to the consumers (internal consumers like its\
      \ employees or B2B customers or B2C customers). \nA service catalog typically\
      \ includes name, description and time period that is valid for. It will have\
      \ a list of ServiceCandidate catalog items. A ServiceCandidate is an entity\
      \ that makes a ServiceSpecification available to a catalog.\nA ServiceCandidate\
      \ and its associated ServiceSpecification may be \"published\" - made visible\
      \ -in any number of ServiceCatalogs, or in none."
    example:
      '@baseType': "@baseType"
      lifecycleStatus: "lifecycleStatus"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      description: "description"
      id: "id"
      href: "href"
      '@schemaLocation': "@schemaLocation"
      version: "version"
  ServiceCatalog_Create:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        description: "Name of the catalog"
      description:
        type: "string"
        description: "Description of this catalog"
      '@type':
        type: "string"
        description: "Indicates the (class) type of catalog. For service catalogs,\
          \ this will be 'ServiceCatalog'."
      '@schemaLocation':
        type: "string"
        description: "This field provides a link to the schema describing this REST\
          \ resource"
      '@baseType':
        type: "string"
        description: "Indicates<b> </b>the base (class) type of this REST resource"
      version:
        type: "string"
        description: "Catalog version"
      validFor:
        description: "The period for which the catalog is valid"
        $ref: "#/definitions/TimePeriod"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status"
    description: "The root entity for service catalog management.\nA service catalog\
      \ is a group of service specifications made available through service candidates\
      \ that an organization provides to the consumers (internal consumers like its\
      \ employees or B2B customers or B2C customers). \nA service catalog typically\
      \ includes name, description and time period that is valid for. It will have\
      \ a list of ServiceCandidate catalog items. A ServiceCandidate is an entity\
      \ that makes a ServiceSpecification available to a catalog.\nA ServiceCandidate\
      \ and its associated ServiceSpecification may be \"published\" - made visible\
      \ -in any number of ServiceCatalogs, or in none.\nSkipped properties: id,href"
    example:
      '@baseType': "@baseType"
      lifecycleStatus: "lifecycleStatus"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      description: "description"
      '@schemaLocation': "@schemaLocation"
      version: "version"
  ServiceCatalog_Update:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Name of the catalog"
      description:
        type: "string"
        description: "Description of this catalog"
      '@schemaLocation':
        type: "string"
        description: "This field provides a link to the schema describing this REST\
          \ resource"
      '@baseType':
        type: "string"
        description: "Indicates<b> </b>the base (class) type of this REST resource"
      version:
        type: "string"
        description: "Catalog version"
      validFor:
        description: "The period for which the catalog is valid"
        $ref: "#/definitions/TimePeriod"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status"
    description: "The root entity for service catalog management.\nA service catalog\
      \ is a group of service specifications made available through service candidates\
      \ that an organization provides to the consumers (internal consumers like its\
      \ employees or B2B customers or B2C customers). \nA service catalog typically\
      \ includes name, description and time period that is valid for. It will have\
      \ a list of ServiceCandidate catalog items. A ServiceCandidate is an entity\
      \ that makes a ServiceSpecification available to a catalog.\nA ServiceCandidate\
      \ and its associated ServiceSpecification may be \"published\" - made visible\
      \ -in any number of ServiceCatalogs, or in none.\nSkipped properties: id,href,id,href,@type,lastUpdate"
    example:
      '@baseType': "@baseType"
      lifecycleStatus: "lifecycleStatus"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      description: "description"
      '@schemaLocation': "@schemaLocation"
      version: "version"
  ServiceCategory:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of the category"
      href:
        type: "string"
        description: "Hyperlink reference to the category"
      name:
        type: "string"
        description: "Name of the category"
      description:
        type: "string"
        description: "Description of the category"
      '@type':
        type: "string"
        description: "The (class) type of this category"
      '@schemalLocation':
        type: "string"
        description: "This field provides a link to the schema describing this REST\
          \ resource"
      '@baseType':
        type: "string"
        description: "Immediate base class type of this category"
      version:
        type: "string"
        description: "Category version"
      validFor:
        description: "The period for which the category is valid"
        $ref: "#/definitions/TimePeriod"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update"
      parentId:
        type: "string"
        description: "Unique identifier of the parent category"
      isRoot:
        type: "boolean"
        description: "If true, this Boolean indicates that the category is a root\
          \ of categories"
      relatedParty:
        type: "array"
        items:
          $ref: "#/definitions/RelatedPartyRef"
      serviceCandidate:
        type: "array"
        items:
          $ref: "#/definitions/ServiceCandidateRef"
      category:
        type: "array"
        items:
          $ref: "#/definitions/CategoryRef"
    description: "The (service) category resource is used to group service candidates\
      \ in logical containers. Categories can contain other categories."
    example:
      lifecycleStatus: "lifecycleStatus"
      isRoot: true
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      serviceCandidate:
      - '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        version: "version"
      - '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        version: "version"
      '@type': "@type"
      description: "description"
      relatedParty:
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
      version: "version"
      parentId: "parentId"
      '@baseType': "@baseType"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      '@schemalLocation': "@schemalLocation"
      id: "id"
      href: "href"
      category:
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
  ServiceCategory_Create:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        description: "Name of the category"
      description:
        type: "string"
        description: "Description of the category"
      '@type':
        type: "string"
        description: "The (class) type of this category"
      '@schemalLocation':
        type: "string"
        description: "This field provides a link to the schema describing this REST\
          \ resource"
      '@baseType':
        type: "string"
        description: "Immediate base class type of this category"
      version:
        type: "string"
        description: "Category version"
      validFor:
        description: "The period for which the category is valid"
        $ref: "#/definitions/TimePeriod"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update"
      parentId:
        type: "string"
        description: "Unique identifier of the parent category"
      isRoot:
        type: "boolean"
        description: "If true, this Boolean indicates that the category is a root\
          \ of categories"
      relatedParty:
        type: "array"
        items:
          $ref: "#/definitions/RelatedPartyRef"
      serviceCandidate:
        type: "array"
        items:
          $ref: "#/definitions/ServiceCandidateRef"
      category:
        type: "array"
        items:
          $ref: "#/definitions/CategoryRef"
    description: "The (service) category resource is used to group service candidates\
      \ in logical containers. Categories can contain other categories.\nSkipped properties:\
      \ id,href"
    example:
      lifecycleStatus: "lifecycleStatus"
      isRoot: true
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      serviceCandidate:
      - '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        version: "version"
      - '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        version: "version"
      '@type': "@type"
      description: "description"
      relatedParty:
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
      - role: "role"
        validFor:
          startDateTime: "2000-01-23T04:56:07.000+00:00"
          endDateTime: "2000-01-23T04:56:07.000+00:00"
        name: "name"
        id: "id"
        href: "href"
      version: "version"
      parentId: "parentId"
      '@baseType': "@baseType"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      '@schemalLocation': "@schemalLocation"
      category:
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
...
  TimePeriod:
    type: "object"
    properties:
      startDateTime:
        type: "string"
        format: "date-time"
        description: "An instant of time, starting at the TimePeriod"
      endDateTime:
        type: "string"
        format: "date-time"
        description: "An instant of time, ending at the TimePeriod."
    description: "A base / value business entity used to represent a period of time\
      \ between two timepoints."
    example:
      startDateTime: "2000-01-23T04:56:07.000+00:00"
      endDateTime: "2000-01-23T04:56:07.000+00:00"
  RelatedPartyRef:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of a related party"
      href:
        type: "string"
        description: "Reference of the related party, could be a party reference or\
          \ a party role reference"
      role:
        type: "string"
        description: "Role of the related party."
      name:
        type: "string"
        description: "Name of the related party"
      validFor:
        description: "Validity period of the related party"
        $ref: "#/definitions/TimePeriod"
    description: "RelatedParty reference. A related party defines party or party role\
      \ linked to a specific entity."
    example:
      role: "role"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      id: "id"
      href: "href"
...
  Error:
    required:
    - "code"
    - "message"
    properties:
      code:
        type: "integer"
        description: "An integer coding the error type. This is given to caller so\
          \ he can translate them if required."
      message:
        type: "string"
        description: "A short localized string that describes the error."
      description:
        type: "string"
        description: "(optional) A long localized error description if needed. It\
          \ can contain precise information about which parameter is missing, or what\
          \ are the identifier acceptable values."
      infoURL:
        type: "string"
        description: "(optional) A URL to online documentation that provides more\
          \ information about the error."